refined move list and move card
